requirements.txt
numpy==2.0.1
python-telegram-bot==21.4
nltk==3.8.2
keras==3.5.0
tensorflow==2.17.0
keras==3.5.0


